Move and update Windows build for Qt6
